refactor: split queries constants
