+++ src/constants.mjs
@@ -6,7 +6,3 @@ export const DOC_NODE_VERSION = process.version;
 // This is the Node.js CHANGELOG to be consumed to generate a list of all major Node.js versions
 export const DOC_NODE_CHANGELOG_URL =
   'https://raw.githubusercontent.com/nodejs/node/HEAD/CHANGELOG.md';
-
-// This is the perma-link within the API docs that reference the Stability Index
-export const DOC_API_STABILITY_SECTION_REF_URL =
-  'documentation.html#stability-index';
