+++ src/generators/legacy-html/utils/buildContent.mjs
@@ -6,7 +6,7 @@ import { SKIP, visit } from 'unist-util-visit';
 
 import buildExtraContent from './buildExtraContent.mjs';
 
-import createQueries from '../../../queries.mjs';
+import createQueries from '../../../utils/queries/index.mjs';
 
 import { DOC_NODE_BLOB_BASE_URL } from '../constants.mjs';
