+++ src/utils/queries/constants.mjs
@@ -1,0 +1,5 @@
+'use strict';
+
+// This is the perma-link within the API docs that reference the Stability Index
+export const DOC_API_STABILITY_SECTION_REF_URL =
+  'documentation.html#stability-index';
