+++ src/utils/tests/queries.test.mjs
@@ -1,6 +1,6 @@
 import { strictEqual, deepStrictEqual } from 'node:assert';
 import { describe, it } from 'node:test';
-import createQueries from '../queries.mjs';
+import createQueries from '../queries/index.mjs';
 
 describe('createQueries', () => {
   it('should add YAML metadata correctly', () => {
